backend/requirements.txt
torch 
transformers 
uvicorn
fastapi
pymongo
python-dotenv